refactor: 규칙에 따라 디렉토리 구조 리팩토링 (only claude)
