+++ src/main/kotlin/com/sight/config/DiscordConfig.kt
@@ -1,4 +1,4 @@
-package com.sight.discord.config
+package com.sight.config
 
 import net.dv8tion.jda.api.JDA
 import net.dv8tion.jda.api.JDABuilder
