+++ src/main/kotlin/com/sight/core/discord/JdaDiscordApiAdapter.kt
@@ -1,6 +1,6 @@
-package com.sight.discord.adapter
+package com.sight.core.discord
 
-import com.sight.discord.model.DiscordRole
+import com.sight.domain.discord.DiscordRole
 import net.dv8tion.jda.api.JDA
 import net.dv8tion.jda.api.entities.Role
 import org.slf4j.LoggerFactory
