+++ src/main/kotlin/com/sight/domain/discord/DiscordRole.kt
@@ -1,4 +1,4 @@
-package com.sight.discord.model
+package com.sight.domain.discord
 
 enum class DiscordRole {
     MEMBER,
